pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gradient-selection"
version = "0.1.0"
description = "Gradient study selection pipeline"
requires-python = ">=3.9"
dependencies = [
  "google-cloud-bigquery>=3.0.0",
  "openai>=1.0.0",
  "pandas>=2.0.0",
  "pyyaml>=6.0",
  "python-dotenv>=1.0.0",
  "typer>=0.12.0",
]

[project.scripts]
gradient-selection = "gradient_selection.cli:app"

[tool.setuptools.packages.find]
where = ["src"]
